Add keys to Card component
When the state of App was changed by clicking next or previous, the
child Card component retained its state. So if it was flipped, then the
next card to be displayed would also be flipped. To reset the Card
component after clicking previous or next, modifying the key would
remount the Card component.